Build wheel archive and serve binaries
This is a start of an experimental feature.
The aim is to compile wheel archive on the mirror
from the PyPI archive.

For exemple,
pip install --use-wheel gevent is built only
once on the mirrored, and servers that need
to install gevent did not need a compiler anymore.
The mirror and the server must use the same architecture
to get this working.

Of course, this is disable by default.