allow flexible configuration of authentication
This commit makes it possible to flexibly configure authentication to the
various parts of the application.

It does this by allowing authenticators to be defined under the
pyshop.auth.methods config keys. pyshop.auth.methods.<name>.type will mean
that if you use <name> in the config as the auth method, the authentication
described by that method will be used. This commit provides  basic,
session, and remote_user mechanisms.

You can also define multiple authenticators of the same type. This is mostly
useful if you want to use remote_user auth multiple times.

The sample config has been updated with default auth mechamisms for basic,
session and remote_user that should be sufficient for most users.

This commit does affect the pyshop.cookie_key config parameter. That key must
now be configured by setting pyshop.auth_methods.<name>.secret on an
auth_method of type session. pyshop.cookie_key is dropped in favor of that.
I've tried to mitigate the impact of that by setting secret on all session
auth methods to pyshop.cookie_key if it does happen to be set. This should
preserve backwards compatibility.

To actually configure part of pyshop to use an auth mechanism, you set
pyshop.auth.<component> to the name of a method that's been defined. e.g. to
set the web-interface to use auth from Apache, you can set pyshop.auth.web to
be remote_user.

The components that can have auth configured are web, simple, and upload. web
is the web-ui, simple is the interface that pip talks to, and upload is what
distutils upload uses for uploading packages.

The split between simple and upload is necessary because while pip can be made
to perform interesting auth (specifically, it can be told to use a client
certificate), distutils cannot. It's limited to just HTTP basic :(