allow configuration of the endpoint for uploading
It's difficult to use different auth for upload and download if they live at
the same URL endpoint. This commit makes the upload endpoint configurable via
the pyshop.upload.route key. If left unset pyshop will continue to use /simple
as the upload endpoint.

Using this key makes it easier to instruct a webserver what type of
authentication to do in the case where different authentication on the upload
and download endpoints are desired.